% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code}
\alias{get_code}
\title{Title}
\usage{
get_code(
  x,
  x_name = NULL,
  filters = NULL,
  columns = ".value",
  rows = NULL,
  values = NULL
)
}
\arguments{
\item{x}{}

\item{filters}{}

\item{columns, rows}{Character vector of column names, one of which must 
include the special `".value"`}

\item{values}{}
}
\description{
Title
}
