% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-glue_inject.R
\name{glue_inject}
\alias{glue_inject}
\title{Insert expressions into a string using rlang injection syntax}
\usage{
glue_inject(..., .open = "<<", .close = ">>", .envir = parent.frame())
}
\arguments{
\item{..., .open, .close, .envir}{Passed to [glue::glue()]}
}
\value{
A character vector
}
\description{
Insert expressions into a string using rlang injection syntax
}
\examples{
a <- expr(1 + 1)
glue_inject("c(<< !!a >>)")
glue_inject("<< c(!!!letters[1:3]) >>")
}
