% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotr-package.R
\docType{package}
\name{pivotr-package}
\alias{pivotr}
\alias{pivotr-package}
\title{pivotr: Excel's PivotTables implemented in shiny and dplyr/tidyr}
\description{
Perform data manipulations using a UI heavily inspired by Excel's PivotTables, then copy the code for those manipulations back into R.
}
\author{
\strong{Maintainer}: Jacob Scott \email{jscott2718@gmail.com}

}
\keyword{internal}
