% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{pivotR}
\alias{pivotR}
\title{Run the pivotR app}
\usage{
pivotR(input_raw, ...)
}
\arguments{
\item{input_raw}{A data frame}

\item{...}{Further arguments passed to pivotR}
}
\value{
A Shiny gadget
}
\description{
Run the pivotR app
}
\examples{
#pivotR(mtcars)
}
