% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Computes grouped aggregates}
\usage{
cube(df, groups = NULL, ..., .totals = "Total")
}
\arguments{
\item{df}{A dataframe}

\item{groups}{variables names}

\item{...}{calculations passed to `dplyr::summarize(...)`}

\item{.totals}{A string of how output totals are labeled in the rows and cols}
}
\description{
Calculates aggregates functions for groups of variables
   and all their combinations
}
\examples{
cube(mtcars, c(cyl, am), Avg = mean(mpg))
}
