% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_pivot_code.R
\name{construct_pivot_code}
\alias{construct_pivot_code}
\title{Title}
\usage{
construct_pivot_code(
  x,
  x_name = NULL,
  filters = NULL,
  columns = ".measure",
  rows = NULL,
  values = NULL
)
}
\arguments{
\item{x}{}

\item{filters}{}

\item{columns, rows}{Character vector of column names, one of which must 
include the special `".measure"`}

\item{values}{}
}
\description{
Title
}
