% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotr-package.R
\docType{package}
\name{pivotr-package}
\alias{pivotr}
\alias{pivotr-package}
\title{pivotr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
