% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotr.R
\name{pivotr}
\alias{pivotr}
\title{Run the {pivotr} app}
\usage{
pivotr()
}
\value{
An object that represents the app
}
\description{
This brings up a GUI similar to Excel's PivotTable interface. Data 
manipulations will be performed and the dplyr/tidyr code used will be
displayed.
}
