% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotr.R
\name{pivotr}
\alias{pivotr}
\title{Run the {pivotr} app}
\usage{
pivotr()
}
\value{
An object that represents the app. Printing the object or passing it 
 to [shiny::runApp()] will run the app.
}
\description{
This brings up a GUI similar to Excel's PivotTable interface. The 
dplyr/tidyr code used to achieve the data transformations can be copied
for modification/programmatic use.
}
